Add onPause and onResume
What?
If user is putting app on background, app will stop ending calls. If
app is in foreground, ending calls is resumed. If queue is still open,
user is notified of this with toast message.

Why?
If user does not close app with back button, but expects app to be
closed, it is good to stop ending calls. People are still
accepted, if queue is open.